updated rollup file with api req from workout file
