+++ rollup.config.js
@@ -4,6 +4,7 @@ import typescript from '@rollup/plugin-typescript';
 import dts from 'rollup-plugin-dts';
 import { terser } from 'rollup-plugin-terser';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
+import copy from 'rollup-plugin-copy';
 
 const packageJson = require('./package.json');
 
@@ -28,6 +29,13 @@ export default [
       commonjs(),
       typescript({ tsconfig: './tsconfig.json' }),
       terser(),
+      // this is for packagejson file for the config file to import
+      copy({
+        targets: [
+          { src: 'package.json', dest: 'dist/esm' }, // Adjust the destination directory if needed
+          { src: 'package.json', dest: 'dist/cjs' }, // Adjust the destination directory if needed
+        ],
+      }),
     ],
     external: ['react', 'react-dom', 'styled-components'],
   },
